Fast forward to 0.3.5

+++ publish.sh
@@ -6,23 +6,18 @@ IFS=$'\n\t,'
 usage(){
     echo "[USAGE] $0 [-h] [-p] <test/master>" 1>&2
     echo
-    echo "Help us publish msiempy on the internet."
+    echo "Help us publish msiempy on the internet. Unix only. "
     echo
     echo -e "\t-h\tPrint this help message."
-    echo -e "\t-p\t<test/master>\tPush the technical documentation, publish to PyPi and Git tag versions. "
+    echo -e "\t-p\t<test/master>\tPublish to PyPi and Git tag versions. "
     echo -e "\t\tInstall the requirements.txt."
     echo
     echo -e "\t\t'test' keyword :"
-    echo -e "\t\t\t- Ask to format the code with black and commit"
-    echo -e "\t\t\t- Publish docs to https://mfesiem.github.io/docs/test/msiempy/"
     echo -e "\t\t\t- Publish module to https://test.pypi.org/project/msiempy/"
     echo
     echo -e "\t\t'master' keyword :"
-    echo -e "\t\t\t- Ask to format the code with black and commit"
-    echo -e "\t\t\t- Print git log since last tag"
-    echo -e "\t\t\t- Publish docs to https://mfesiem.github.io/docs/msiempy/"
     echo -e "\t\t\t- Publish module to https://pypi.org/project/msiempy/"
-    echo -e "\t\t\t- Ask if should tag the version and interactively ask you a message with vi."
+    echo -e "\t\t\t- Ask to tag the version and interactively ask you a message with vi."
     echo -e "\t\t\t- Note that you'll still need to create the realease from github"
     echo
     exit -1
@@ -40,8 +35,16 @@ while getopts ":hp:" arg; do
             python3 setup.py install
             
             # Figuring version
-            version="$(grep __version__ ./msiempy/__version__.py | cut -d '"' -f 2)"
-        
+            version="$(python3 setup.py -V)"
+            
+            # Generate the index.md file and push it
+            ./build_docs -i -f
+            cp -f "./docs/${version}/index.md" "./docs/index.md"
+            
+            git add ./docs
+            git commit -m "Generate docs/${version}/index.md"
+            git push
+
             # Checking keyword
             keyword=${OPTARG}
 
@@ -58,57 +61,6 @@ while getopts ":hp:" arg; do
                     exit -1
                 fi
             fi
-
-            # Ask to format code and push it
-            echo "[QUESTION] Do you want to format code with black ?"
-            read REPLY < /dev/tty
-            if [ "$REPLY" = "y" ] || [ "$REPLY" = "yes" ] || [ "$REPLY" = "Y" ] || [ "$REPLY" = "Yes" ]; then
-                black "msiempy/"
-                git add "msiempy"
-                git commit -m "Format code"
-                git push
-            fi
-
-            
-            if [ "$keyword" = "master" ]; then
-                last_tag=`git tag -l | tail -1`
-                if [[ ! -z ${last_tag} ]]; then
-                    git config pager.branch false
-                    compare=`git log ${last_tag}.. --pretty=oneline`
-                else
-                    compare='No last tag to compare with'
-                fi
-                echo "[INFO] Git log since last tag: "
-                echo ${compare}
-            
-            fi
-
-            # Generating diagrams
-            echo "[RUNNING] pyreverse -s 1 -f PUB_ONLY -o png -m y msiempy"
-            pyreverse -s 1 -f PUB_ONLY -o png -m y msiempy
-
-            # Cloning or pulling changes from the documentation
-            if [[ ! -d mfesiem.github.io ]]; then
-                git clone https://github.com/mfesiem/mfesiem.github.io
-            else
-                cd mfesiem.github.io && git pull --quiet && cd ..
-            fi
-            
-            # Generating documentation
-            echo "[RUNNING] pdoc"
-            rm -rf ./${docs_folder}/msiempy/
-            python3 -m pdoc msiempy --output-dir ./${docs_folder} --html --force --template-dir ./.pdoc_templates
-
-
-            mv ./classes.png ./${docs_folder}/msiempy
-            mv ./packages.png ./${docs_folder}/msiempy
-            
-            # Pushing docs
-            echo "[RUNNING] pushing docs"
-            cd mfesiem.github.io && git add . && git commit -m "Generate ${keyword} docs $(date)" --quiet && git push origin master --quiet
-            cd ..
-            
-            echo "[SUCCESS] Documentation at : https://${docs_folder}/msiempy/"
 
             # Building module
             echo "[RUNNING] building"
@@ -127,31 +79,32 @@ while getopts ":hp:" arg; do
             echo "[SUCCESS] Module published at : https://${repository_url}/project/msiempy/"
 
             # Ask to Tag ?
-            read -p "[QUESTION] Do you want to tag this version '${version}'? You'll be asked to write the tag message. [y/n]" -n 1 -r
-            echo    # (optional) move to a new line
-            if [[ $REPLY =~ ^[Yy]$ ]]; then
-
-                # Tag
-                touch ./tmp_tag.txt
-                echo "msiempy ${version}" > ./tmp_tag.txt
-                echo >> ./tmp_tag.txt
-                echo "New features: " >> ./tmp_tag.txt
-                echo >> ./tmp_tag.txt
-                echo "Fixes:" >> ./tmp_tag.txt
-                vi ./tmp_tag.txt
-                tag_msg=`cat ./tmp_tag.txt`
-                echo "${tag_msg}"
-                read -p "[QUESTION] Are you sure, tag this version with the message? [y/n]" -n 1 -r
+            if [ "$keyword" = "master" ]; then
+                read -p "[QUESTION] Do you want to tag this version '${version}'? You'll be asked to write the tag message. [y/n]" -n 1 -r
                 echo    # (optional) move to a new line
                 if [[ $REPLY =~ ^[Yy]$ ]]; then
-                    echo "[RUNNING] pushing tags"
-                    git tag -a ${version} -F ./tmp_tag.txt && git push --tags
-                    echo "[SUCCESS] msiempy ${version} tagged and pushed to https://github.com/mfesiem/msiempy/tags"
-                    echo "[INFO] Note that you'll still need to create the realease from github"
-                fi
-                rm ./tmp_tag.txt
-            fi
 
+                    # Tag
+                    touch ./tmp_tag.txt
+                    echo "msiempy ${version}" > ./tmp_tag.txt
+                    echo >> ./tmp_tag.txt
+                    echo "New features: " >> ./tmp_tag.txt
+                    echo >> ./tmp_tag.txt
+                    echo "Fixes:" >> ./tmp_tag.txt
+                    vi ./tmp_tag.txt
+                    tag_msg=`cat ./tmp_tag.txt`
+                    echo "${tag_msg}"
+                    read -p "[QUESTION] Are you sure, tag this version with the message? [y/n]" -n 1 -r
+                    echo    # (optional) move to a new line
+                    if [[ $REPLY =~ ^[Yy]$ ]]; then
+                        echo "[RUNNING] pushing tags"
+                        git tag -a ${version} -F ./tmp_tag.txt && git push --tags
+                        echo "[SUCCESS] msiempy ${version} tagged and pushed to https://github.com/mfesiem/msiempy/tags"
+                        echo "[INFO] Note that you'll still need to create the realease from github"
+                    fi
+                    rm ./tmp_tag.txt
+                fi
+            fi
             ;;
 
         *)
@@ -162,4 +115,4 @@ while getopts ":hp:" arg; do
     esac
 done
 shift $((OPTIND-1))
-echo "[END] usage: $0 [-h] [-p <test/master>]"
+echo "[END]"
